src/pyproject.toml
[tool.poetry]
name = "imse"
version = "0.1.0"
description = "Light weight image search tool."
authors = ["benjamindev <benjamin.decharmoy@gmail.com>"]

packages = [
    {include = "imse"}
]

[tool.poetry.dependencies]
python = "^3.7"
loguru = "^0.5.1"
simplekml = "^1.3.5"
tqdm = "^4.48.0"
pillow = "^7.2.0"

[tool.poetry.scripts]
pix2map = 'imse.tools.cli:pix2map'

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
